barcode/requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
Pillow==9.1.0
python-barcode==0.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
